refactor: moved configs back to the init factory and removed not needed init method
